Binarycoin v0.6.4.1 Update
This update features changes mainly to the UI, such as the addition of
three new buttons: CryptoCoinTalk, CCEX Exchange and a Block Explorer
button (Disabled).
This update is mainly aimed to update the list of providers and
supporters of Binary and does not feature and any Major Functionality
changes otherwise.
Scheduled next release of wallet v0.6.4.2 will focus on more
functionality improvements.